Book_Summarizer-RAG/requirements.txt
groq
langchain
faiss-cpu
python-dotenv
pypdfium2
tqdm
sentence-transformers
fpdf
pytesseract
Pillow